% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-em.R
\name{em.control}
\alias{em.control}
\title{Create a control structure for running EM algorithms}
\usage{
em.control(ctrl = NULL, ...)
}
\arguments{
\item{nplot}{how often to plot wages of comparaison to model0}

\item{ncat}{how often to log information}

\item{maxiter}{maximum number of iterations}

\item{model_var}{whether to allow for flexible variance (default is TRUE)}

\item{est_Amb}{TBD}

\item{cstr_type}{defines the type of constraints on the means for estimation}

\item{cstr_val}{TBD}

\item{tol}{tolerance for stopping the EM}

\item{tau}{posterior likelihood to use instead of computing them using initial parameters}

\item{model0}{model to compare estimation to when plotting}

\item{presp=1e-9}{TBD}

\item{rel_weight}{setting a particular weight for stayers versus movers (default is 1)}

\item{est_rho}{vector of TRUE/FALSE that states which rho should be estimated}

\item{rho_in_diff}{whether to estimate rho in level of in differences}

\item{dprior}{Dirichlet prior for proportions (default 1.01)}

\item{nfirms_to_update=10, }{number firms to update in probabilistic approach}

\item{proba_include}{what terms to include in the liklihood for probabilistic approach (default to = c(1,1,1,1))}

\item{check_lik}{whether to check the likelihood at each updating parameter}

\item{stochastic}{whether to use stochastic EM instead of straight EM}

\item{fixb}{when TRUE, imposes fixed interactions in different time periods}

\item{fixm}{when TRUE, levels are not updated, only variances and proportions}

\item{deps=1e-50}{TBD}

\item{file_backup_prefix}{TBD}

\item{sd_floor=}{floor imposed on all standard deviations (default is 1e-10)}

\item{posterior_reg}{term added to posterior probablities (this is to deal with numerical issues, default is 1e-9)}

\item{textapp}{text to show in logging}

\item{sdata_subsample}{share of the stayers to use in estimation}

\item{sdata_subredraw}{whether to redraw the subsample of stayers}

\item{vdec_sim_size}{size to use in simulation}

\item{stayer_weight}{weight attributed to stayers in joint estimation}

\item{est_rep=}{number of starting values for EM}

\item{est_nbest}{number of best starting values to select from using connectedness}
}
\description{
Create a control structure for running EM algorithms
}
