% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three-sided-utils.R
\name{estimation.threeSided.model}
\alias{estimation.threeSided.model}
\title{Mixture model estimation of three sided model}
\usage{
estimation.threeSided.model(model_test, ad_employee_em, ctrl)
}
\description{
Mixture model estimation of three sided model
}
