% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2-trace.R
\name{get.largest.conset.fid}
\alias{get.largest.conset.fid}
\title{some work on trace formula
Extracts the largest connected set from data using f1,f2
movers}
\usage{
get.largest.conset.fid(jdata)
}
\description{
some work on trace formula
Extracts the largest connected set from data using f1,f2
movers
}
