% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2-trace.R
\name{get.largest.leaveoutset.fid}
\alias{get.largest.leaveoutset.fid}
\title{Extracts the largest leave-out connected set from data using f1,f2
movers}
\usage{
get.largest.leaveoutset.fid(jdata)
}
\description{
Extracts the largest leave-out connected set from data using f1,f2
movers
}
