% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firm-clustering.R
\name{grouping.append}
\alias{grouping.append}
\title{Append result of a grouping to a data-set}
\usage{
grouping.append(sim, clus, drop = F, sort = T)
}
\arguments{
\item{sim}{a data set with jdata and sdata}

\item{clus}{a clustering result from the classify function}

\item{drop}{if true firms in sim, not in cluster are dropped}
}
\description{
Append result of a grouping to a data-set
}
