% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firm-clustering.R
\name{grouping.classify.once}
\alias{grouping.classify.once}
\title{clusters firms based on their cross-sectional wage distributions}
\usage{
grouping.classify.once(
  measures,
  k = 10,
  nstart = 1000,
  iter.max = 200,
  step = 20
)
}
\arguments{
\item{measures}{object created using grouping.getMeasures}

\item{k}{number of groups}

\item{nstart}{(default:1000) total number of starting values}

\item{iter.max}{(default:100) max number of step for each repetition}

\item{step}{step size in the repeating}

\item{sdata}{cross sectional data, needs a column j (firm id) and w (log wage)}

\item{Nw}{number of points to use for wage distribution}
}
\description{
clusters firms based on their cross-sectional wage distributions
}
