% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firm-clustering.R
\name{grouping.classify}
\alias{grouping.classify}
\title{clusters firms based on their cross-sectional wage distributions}
\usage{
grouping.classify(
  measures,
  ksupp = ceiling((1:(60)^(1/1.3))^1.3),
  nstart = 1000,
  iter.max = 200,
  stop = FALSE,
  verbose = 1,
  cval = 1
)
}
\arguments{
\item{measures}{specify the type of measures to use (mean and var, quantiles, etc...)}

\item{ksupp}{vector of different number of groups to try}

\item{nstart}{(default:1000) total number of starting values}

\item{iter.max}{(default:100) max nunmber of step for each repetition}

\item{sdata}{cross sectional data, needs a column j (firm id) and w (log wage)}

\item{Nw}{number of points to use for wage distributionsdsd}

\item{M}{you can pass the matrix measurements, requires also weights W (pass on the truth for instance)}
}
\description{
clusters firms based on their cross-sectional wage distributions
}
