% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firm-clustering.R
\name{grouping.computeobj}
\alias{grouping.computeobj}
\title{Compute the objective function of the clustering}
\usage{
grouping.computeobj(grps)
}
\description{
Compute the objective function of the clustering
}
