% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firm-clustering.R
\name{grouping.getMeasures.em}
\alias{grouping.getMeasures.em}
\title{Get the three sided measures for inputs in clustering}
\usage{
grouping.getMeasures.em(
  sim,
  y_var = "y1",
  measure = "ecdf",
  Nw = 40,
  model = NA,
  exclude_movers = FALSE,
  user_fn = NA,
  three_way = T
)
}
\description{
Get the three sided measures for inputs in clustering
}
