% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firm-clustering.R
\name{grouping.getMeasures}
\alias{grouping.getMeasures}
\title{Extract the measurement matrix to be given to the
classification algorithm}
\usage{
grouping.getMeasures(
  sim,
  y_var = "y1",
  measure = "ecdf",
  Nw = 40,
  model = NA,
  exclude_movers = FALSE,
  user_fn = NA,
  three_way = T
)
}
\description{
Extract the measurement matrix to be given to the
classification algorithm
}
