% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firm-clustering.R
\name{grouping.makefiner}
\alias{grouping.makefiner}
\title{Gives their won cluster to firm with many movers}
\usage{
grouping.makefiner(clus, sim, nm = 30)
}
\arguments{
\item{clus}{a clustering result from the classify function}

\item{sim}{a data set with jdata and sdata}
}
\description{
Gives their won cluster to firm with many movers
}
