% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc.R
\name{jdata.prepare}
\alias{jdata.prepare}
\title{Prepare the data for BLM from an employer-employee matched data}
\usage{
jdata.prepare(data)
}
\description{
This relies on the data with the following variables:
 - wid worker id
 - fid firm id
 - y is the log-wage
 - t is time
}
