% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firm-clustering.R
\name{kmeansW.repeat}
\alias{kmeansW.repeat}
\title{internal function that runs Kmean with multiple starting values}
\usage{
kmeansW.repeat(
  x,
  centers,
  weight = rep(1, nrow(x)),
  iter.max = 100,
  nstart = 1,
  step = 20
)
}
\description{
internal function that runs Kmean with multiple starting values
}
