% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-proj.R
\name{lin.proja}
\alias{lin.proja}
\title{Generate a linear projection decomposition for the model
with continuous worker hetergoneity}
\usage{
lin.proja(sdata, y_col = "y", k_col = "k", j_col = "j", do_interacted_reg = 1)
}
\description{
Generate a linear projection decomposition for the model
with continuous worker hetergoneity
}
