% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2-mixt-three-sided-v2.R
\name{m2.get.pk_unc}
\alias{m2.get.pk_unc}
\title{Returns the uconditional type probability in the crossection}
\usage{
m2.get.pk_unc(model, supersample = 0.1)
}
\description{
Returns the uconditional type probability in the crossection
}
