% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2-mixt-three-sided-v2.R
\name{m2.mixt.estimate.all}
\alias{m2.mixt.estimate.all}
\title{Estimates the static mixture model on 2 periods}
\usage{
m2.mixt.estimate.all(sim, nk = 6, ctrl, cl = NA, nbb = 1)
}
\description{
This estimator uses multiple starting values to try to find the global maxima.
}
