% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2-mixt-three-sided-v2.R
\name{m2.mixt.movers}
\alias{m2.mixt.movers}
\title{Estimates the static model parameters for movers}
\usage{
m2.mixt.movers(jdatae, model, ctrl)
}
\description{
Estimates the static model parameters for movers
}
