% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2-mixt-three-sided-v2.R
\name{m2.mixt.new}
\alias{m2.mixt.new}
\title{create a random model for EM with three sided
endogenous mobility with multinomial pr}
\usage{
m2.mixt.new(nk, nf, nb, fixb = F, stationary = F)
}
\description{
create a random model for EM with three sided
endogenous mobility with multinomial pr
}
