% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2-mixt-three-sided-v2.R
\name{m2.mixt.simulate.movers}
\alias{m2.mixt.simulate.movers}
\title{Using the model, simulates a dataset of movers}
\usage{
m2.mixt.simulate.movers(model, NNm = NA)
}
\description{
Using the model, simulates a dataset of movers
}
