% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2-mixt-three-sided-v2.R
\name{m2.mixt.simulate.sim.clust}
\alias{m2.mixt.simulate.sim.clust}
\title{Simulates data (movers and stayers)}
\usage{
m2.mixt.simulate.sim.clust(model, fsize, msize, smult = 1, mmult = 1)
}
\description{
Simulates data (movers and stayers)
}
