% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2-mixt-three-sided-v2.R
\name{m2.mixt.simulate.sim}
\alias{m2.mixt.simulate.sim}
\title{Simulates data (movers and stayers) and attached firms ids. Firms have all same expected size.}
\usage{
m2.mixt.simulate.sim(model, fsize, msize, smult = 1, mmult = 1)
}
\description{
Simulates data (movers and stayers) and attached firms ids. Firms have all same expected size.
}
