% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2-mixt-three-sided-v2.R
\name{m2.mixt.simulate.stayers}
\alias{m2.mixt.simulate.stayers}
\title{Using the model, simulates a dataset of stayers.}
\usage{
m2.mixt.simulate.stayers(model, NNs)
}
\description{
Using the model, simulates a dataset of stayers.
}
