% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2-mixt-three-sided-v2.R
\name{m2.mixt.simulate.stayers.withx}
\alias{m2.mixt.simulate.stayers.withx}
\title{Using the model, simulates a dataset of stayers.}
\usage{
m2.mixt.simulate.stayers.withx(model, NNsx)
}
\description{
Using the model, simulates a dataset of stayers.
}
