% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2-mixt-three-sided-v2.R
\name{m2.mixt.stayers}
\alias{m2.mixt.stayers}
\title{use the marginal distributions to extract type distributions
within each cluster and observable characteristics}
\usage{
m2.mixt.stayers(sdata, model, ctrl)
}
\description{
use the marginal distributions to extract type distributions
within each cluster and observable characteristics
}
