% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three-sided-utils.R
\name{m2.mixt.transform.data}
\alias{m2.mixt.transform.data}
\title{Data tranformation ( 3d array to 2d array)}
\usage{
m2.mixt.transform.data(model_3d, model_2d_init, data = NA)
}
\description{
Data tranformation ( 3d array to 2d array)
}
