% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three-sided-utils.R
\name{m2.mixt.transform.model}
\alias{m2.mixt.transform.model}
\title{Model transformation for solver (3d array object to 2d array objects)
used in multicore effitiency}
\usage{
m2.mixt.transform.model(model_3d, model_2d_init, data = NA)
}
\description{
Model transformation for solver (3d array object to 2d array objects)
used in multicore effitiency
}
