% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2-mixt-three-sided-v2.R
\name{m2.mixt.vdec}
\alias{m2.mixt.vdec}
\title{Computes the variance decomposition by simulation}
\usage{
m2.mixt.vdec(model, nsim, stayer_share = 1, ydep = "y2")
}
\description{
Computes the variance decomposition by simulation
}
