% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2-mixt-three-sided-v2.R
\name{m2.movers.checkfit}
\alias{m2.movers.checkfit}
\title{check the fit in the movers/stayers using imputed data}
\usage{
m2.movers.checkfit(jdata)
}
\description{
check the fit in the movers/stayers using imputed data
}
