% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2-mixt-three-sided-v2.R
\name{m2.stayers.checkfit}
\alias{m2.stayers.checkfit}
\title{check the fit in the movers/stayers using imputed data}
\usage{
m2.stayers.checkfit(sdata, r1, r4)
}
\description{
check the fit in the movers/stayers using imputed data
}
