% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2-trace.R
\name{m2.trace.new}
\alias{m2.trace.new}
\title{create a model for testing trace estimation}
\usage{
m2.trace.new(nf = 200, nm = 10000, eps_sd = 1.5)
}
\description{
create a model for testing trace estimation
}
