% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mcast}
\alias{mcast}
\title{creates a matrix and fill it using a data.table
in contrast to acast, it creates all rows and cols (even if there is not data)}
\usage{
mcast(dd, val, row, col, dim, fill = NA)
}
\description{
creates a matrix and fill it using a data.table
in contrast to acast, it creates all rows and cols (even if there is not data)
}
