% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{model.connectiveness}
\alias{model.connectiveness}
\title{Computes graph connectedness among the movers
within each type and returns the smalless value}
\usage{
\method{model}{connectiveness}(model, all = FALSE)
}
\description{
Computes graph connectedness among the movers
within each type and returns the smalless value
}
