% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three-sided-utils.R
\name{ModelInitializer}
\alias{ModelInitializer}
\title{Prepare the finction for test run
intialize the model parameters (helper function)}
\usage{
ModelInitializer()
}
\description{
Prepare the finction for test run
intialize the model parameters (helper function)
}
