% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc.R
\name{plot.trquant}
\alias{plot.trquant}
\title{plots the conditional quantile distribution for each transitions (l,l')}
\usage{
\method{plot}{trquant}(res, jdata, jdata2 = NULL)
}
\description{
plots the conditional quantile distribution for each transitions (l,l')
}
