% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataproc.R
\name{plot.vaeffect}
\alias{plot.vaeffect}
\title{we want to look at the effect of
movers on value added.}
\usage{
\method{plot}{vaeffect}()
}
\description{
we want to look at the effect of
movers on value added.
}
