% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataplots.R
\name{plot.wage}
\alias{plot.wage}
\title{plot the mean wage at origin conditional on where it is coming from}
\usage{
\method{plot}{wage}(jdata)
}
\description{
plot the mean wage at origin conditional on where it is coming from
}
