% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three-sided-utils.R
\name{set.solver.controls}
\alias{set.solver.controls}
\title{Set the solver controls}
\usage{
set.solver.controls(
  model_test = model_test,
  n_startValues = 1,
  stayers_sample = 0.1
)
}
\description{
Set the solver controls
}
