% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three-sided-utils.R
\name{Simulate.data.threeSided}
\alias{Simulate.data.threeSided}
\title{Simulate the data consiting of three sided heterogeniety}
\usage{
Simulate.data.threeSided(model_test)
}
\description{
Simulate the data consiting of three sided heterogeniety
}
