% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{spread}
\alias{spread}
\title{this is a utility function to generate
multidimensional arrays - like the spread function in fortran}
\usage{
spread(A, loc, dims)
}
\description{
this is a utility function to generate
multidimensional arrays - like the spread function in fortran
}
