% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/three-sided-utils.R
\name{threeSided.Clustering}
\alias{threeSided.Clustering}
\title{Clustering (Step 1 of the estimation: Seep paper for details)}
\usage{
threeSided.Clustering(
  test_data,
  moments = "ecdf",
  Nw = 20,
  y_var = "y1",
  three_way = "F"
)
}
\description{
Clustering (Step 1 of the estimation: Seep paper for details)
}
