% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vcast}
\alias{vcast}
\title{creates a vector and fill it using a data.table
in contrast to acast, it creates all elements}
\usage{
vcast(dd, val, index, size, fill = NA)
}
\description{
creates a vector and fill it using a data.table
in contrast to acast, it creates all elements
}
